changed some columns to use floats

+++ SQL/Tables/app-physical-table.sql
@@ -5,7 +5,7 @@ GO
 CREATE TABLE [app].[physical](
 	[PhysicalID] [int] IDENTITY(1,1) NOT NULL,
 	[UserID] [int] NOT NULL,
-	[PainLevel] [int] NULL,
+	[PainLevel] [float] NULL,
 	[Headache] [bit] NULL,
 	[Nausea] [bit] NULL,
 	[StomachAche] [bit] NULL,
